requirements.txt
absl-py==2.4.0
cffi==2.0.0
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
flatbuffers==25.12.19
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
mediapipe==0.10.32
numpy==2.4.2
opencv-contrib-python==4.13.0.90
opencv-python==4.13.0.90
packaging==26.0
pillow==12.1.0
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
sounddevice==0.5.5
